# server/specs/openapi.yaml
openapi: 3.1.0
info:
  title: Dinnermate API
  version: 1.0.0
  description: >
    Group restaurant decision-making: create a room, share its code, swipe,
    and see matches. Shareable lists ride along. All non-health routes
    identify the caller via the `X-Dinnermate-User` header (a client-generated
    UUID); there is no authentication beyond it in v1.
servers:
  - url: /
    description: Same-origin (nginx proxies /api to the API container)
...
security:
  - userId: []
paths:
  /api/v1/rooms:
    post:
      tags: [rooms]
      operationId: createRoom
      summary: Create a room
      description: >
        Validates params, fetches and filters a restaurant deck from the
        provider, and snapshots it. The creator is NOT auto-joined; call
        join afterwards.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: Room created with its snapshotted deck
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRoomResponse'
        '400':
          $ref: '#/components/responses/MissingUser'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '502':
          $ref: '#/components/responses/BadGateway'
...
components:
  securitySchemes:
    userId:
      type: apiKey
      in: header
      name: X-Dinnermate-User
      description: >
        Client-generated UUID identifying the caller. Absent or non-UUID
        values yield 400 with code MISSING_USER.
...
  responses:
    MissingUser:
      description: X-Dinnermate-User header absent or not a UUID (code MISSING_USER)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UnprocessableEntity:
      description: Invalid params or unknown restaurant (code INVALID_PARAMS or UNKNOWN_RESTAURANT)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadGateway:
      description: Restaurant provider unavailable or returned garbage (code PROVIDER_UNAVAILABLE)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ShareCode:
      type: string
      pattern: '^[ABCDEFGHJKMNPQRSTUVWXYZ23456789]{6}$'
      examples: [TQ4K7M]
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - MISSING_USER
                - ROOM_NOT_FOUND
                - LIST_NOT_FOUND
                - NOT_IN_ROOM
                - NOT_LIST_MEMBER
                - ALREADY_SWIPED
                - UNKNOWN_RESTAURANT
                - OWNER_CANNOT_LEAVE
                - INVALID_PARAMS
                - PROVIDER_UNAVAILABLE
                - INTERNAL
            message:
              type: string
...
    CreateRoomRequest:
      type: object
      required:
        - location_label
        - lat
        - lng
        - radius_m
        - cuisines
        - price_min
        - price_max
        - min_rating
      properties:
        name:
          type: [string, 'null']
        location_label:
          type: string
        lat:
          type: number
          format: double
          minimum: -90
          maximum: 90
        lng:
          type: number
          format: double
          minimum: -180
          maximum: 180
        radius_m:
          type: integer
          minimum: 100
          maximum: 40000
        cuisines:
          type: array
          items:
            type: string
          description: Lowercase cuisine tags; empty array means all cuisines.
        price_min:
          type: integer
          minimum: 1
          maximum: 4
        price_max:
          type: integer
          minimum: 1
          maximum: 4
          description: Must be >= price_min.
        min_rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
    Room:
      type: object
      description: Room with its search params flattened (same field names as CreateRoomRequest).
      required:
        - id
        - code
        - name
        - location_label
        - lat
        - lng
        - radius_m
        - cuisines
        - price_min
        - price_max
        - min_rating
        - created_at
      properties:
        id:
          type: string
          format: uuid
        code:
          $ref: '#/components/schemas/ShareCode'
        name:
          type: [string, 'null']
        location_label:
          type: string
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        radius_m:
          type: integer
        cuisines:
          type: array
          items:
            type: string
        price_min:
          type: integer
        price_max:
          type: integer
        min_rating:
          type: number
          format: float
        created_at:
          type: string
          format: date-time
    HoursPeriod:
      type: object
      description: >
        Weekly opening span in the restaurant's local time. A close earlier
        than open means the span crosses midnight into the next day.
      required: [day, open, close]
      properties:
        day:
          type: integer
          minimum: 0
          maximum: 6
          description: 0 = Sunday .. 6 = Saturday
        open:
          type: string
          pattern: '^\d{2}:\d{2}$'
          examples: ['11:00']
        close:
          type: string
          pattern: '^\d{2}:\d{2}$'
          examples: ['22:00']
    Restaurant:
      type: object
      required:
        - id
        - name
        - cuisine
        - price_level
        - rating
        - rating_count
        - address
        - photo_url
        - lat
        - lng
        - hours
        - utc_offset_minutes
      properties:
        id:
          type: string
          description: Provider-scoped id, e.g. "seed-001" or a Google place id.
        name:
          type: string
        cuisine:
          type: string
          description: Lowercase single tag, e.g. "thai".
        price_level:
          type: integer
          minimum: 1
          maximum: 4
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
        rating_count:
          type: integer
          minimum: 0
        address:
          type: string
        photo_url:
          type: [string, 'null']
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        hours:
          oneOf:
            - type: array
              items:
                $ref: '#/components/schemas/HoursPeriod'
            - type: 'null'
          description: Weekly opening hours; null when unknown.
        utc_offset_minutes:
          type: [integer, 'null']
          description: Restaurant-local UTC offset in minutes; null when unknown.
...
    CreateRoomResponse:
      type: object
      required: [room, deck]
      properties:
        room:
          $ref: '#/components/schemas/Room'
        deck:
          type: array
          items:
            $ref: '#/components/schemas/Restaurant'
...